mart_agent/pyproject.toml
[tool.poetry]
name = "mart-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langgraph = "^0.2.60"
langchain-core = "^0.3.28"
langchain-google-genai = "^2.0.7"
langgraph-cli = {extras = ["inmem"], version = "^0.1.65"}
langchain-openai = "^0.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
